Resolve open SonarQube code-quality issues
Address the outstanding SonarQube findings across the view layer.

Code fixes:
- ArtistViewController: guard the artist-catalog Optional with ifPresent instead
  of calling get() unconditionally (S3655).
- NavigationController: drop the unused EasyBind import (S1128); make the
  navigation menu list a local variable (S1450); split exportSelectedPlaylists
  into helpers to bring its cognitive complexity under the threshold (S3776).
- PlaylistTreeView: remove the redundant Integer cast on the playlist id (S1905).
- ArtistAlbumListRow: extract the repeated "album-info-secondary" style class
  into a constant (S1192).
- GenreViewController / AlbumViewController: make the grid-backing lists local
  variables (S1450).
- OverlayTracksDrawer: use Math.clamp instead of nested Math.min/Math.max (S6885).

False positives (marked in SonarQube, suppressed in code):
- Defensive null guards on non-null domain getters in AudioItemQueryMatcher,
  AlbumViewController, and GenreViewController (S2589/S2583): mocks and partial
  catalogs can yield nulls at runtime, and a unit test enforces this safety.
- Genre cover-pool / grid-refresh / play helpers that coordinate controller-wide
  state rather than a single grid cell (S3398).
- CSS "unknown pseudo-class" warnings on JavaFX stylesheets (S4659): JavaFX uses
  its own pseudo-classes, so the W3C-CSS rule is excluded for the css files.

Build configuration:
- Fix the CSS issue-exclusion entry to use the mandatory resourceKey file
  pattern; the previous resourcePath key failed the analysis with
  "File pattern is mandatory".
- Gradle 10 readiness: reference test fixtures via project(':') string notation
  instead of the deprecated Project-object dependency notation.

+++ src/main/java/net/transgressoft/musicott/view/custom/PlaylistTreeView.java
@@ -517,7 +517,7 @@ public class PlaylistTreeView extends TreeView<ObservablePlaylist> {
                 var clipboardContent = new ClipboardContent();
                 // Put the serializable integer id on the dragboard; the drop handler
                 // resolves it back to the playlist via the hierarchy.
-                clipboardContent.put(PLAYLIST_DATA_FORMAT, (Integer) getItem().getId());
+                clipboardContent.put(PLAYLIST_DATA_FORMAT, getItem().getId());
                 dragBoard.setContent(clipboardContent);
             }
             event.consume();
